+++ composeApp/src/commonMain/kotlin/com/ribod/kdraw/ui/core/ex/ModiferEx.kt
@@ -7,9 +7,9 @@ import androidx.compose.ui.geometry.Offset
 @Composable
 expect fun Modifier.drawingCanvasModifier(
     offset: Offset,
-    scale: Float,
+    scale: Int,
     changeOffset: (Offset, Boolean) -> Unit,
-    changeScale: (Float) -> Unit,
+    changeScale: (Int) -> Unit,
     changeBoxOffSet: (Offset, Boolean) -> Unit,
-    changeBoxScale: (Float) -> Unit
+    changeBoxScale: (Int) -> Unit
 ): Modifier
